+++ CHANGELOG.md
@@ -101,7 +101,7 @@
 
 ### 🐛 Bug Fixes
 
-* **addon:** add TLS param to MateriaDB KV connect info ([dcd01c1](https://github.com/CleverCloud/clever-tools/commit/dcd01c1ab2b0eb93a9ec9063d21f393216d95b4c))
+* **addon:** add TLS param to Materia KV connect info ([dcd01c1](https://github.com/CleverCloud/clever-tools/commit/dcd01c1ab2b0eb93a9ec9063d21f393216d95b4c))
 * **addon:** allow json format in new list subcommand ([cec0e85](https://github.com/CleverCloud/clever-tools/commit/cec0e8567286e1b56459fbc5e53da91b8ad7765b))
 * **logs:** add details about configuration for specific drains ([55054d4](https://github.com/CleverCloud/clever-tools/commit/55054d430c4b2bcec73cbfbd20fe9e63706c82fe))